requirements.txt
streamlit
onnxruntime==1.17.1
opencv-python-headless
numpy
pillow
requests
